Added Classes for Form Entry Data
There are 3 classes, all currently incomplete, to manage data. One is for the individual data fields. Another is for the initialization of the data from the properties file into an array of objects. Lastly is a data controller designed to prevent the multiple thread (still to come) from submitting the same information repeatedly.